+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Migration notes:
 * Changed: The plugin will now shutdown in case a severe issue prevents loading the config. This includes the case that the config version is invalid. Previously it would treat invalid and missing config versions the same and apply config migrations nevertheless.
 * Changed: Changed/Added a few information/warning messages related to config and language file loading.
 * Changed: Added more information to the message that gets logged when a shopkeeper gets removed for owner inactivity.
+* Changed: Added validation that the unique id of loaded or freshly created shopkeepers is not yet used.
 
 * API: Added ShopkeepersStartupEvent which can be used by plugins to make registrations during Shopkeepers' startup process (eg. to register custom shop types, object types, etc.). This event is marked as deprecated because custom shop types, object types, etc. are not yet officially supported as part of the API. Also, the event is called so early that the plugin (and thereby the API) are not yet fully setup and ready to be used, so this event is only of use for plugins which know what they are doing.
 * API: Removed various API methods from Shopkeeper which simply delegated to the corresponding shop object.